Separate withLock function from Lock model
There's no real reason that function needs to live on the model itself,
and in fact the interface is somewhat cleaner if it doesn't.

(As part of the migration, instead of yielding the lock ID, yield a
function which can renew the lock. That's all we needed it for anyway.)
